Separate file types from factory and definitions.
